Crate jsonrpsee_server

jsonrpsee-server

jsonrpsee-server is a JSON RPC server that supports both HTTP and WebSocket transport.

Re-exports

pub use jsonrpsee_types as types;

pub use tracing;

Modules

helpers

Helpers.

logger

Logger for jsonrpsee servers.

middleware

Various middleware implementations for RPC specific purposes.

Structs

BatchResponseBuilder

Builder to build a BatchResponse.

BoundedSubscriptions

This wraps tokio::sync::Semaphore and is used to limit the number of subscriptions per connection.

BoundedWriter

Bounded writer that allows writing at most max_len bytes.

DisconnectError

Error that may occur during crate::server::MethodSink::send or crate::server::SubscriptionSink::send.

IsUnsubscribed

Represents a subscription until it is unsubscribed.

MethodResponse

Represent the response to a method call.

MethodSink

Sink that is used to send back the result to the server for a specific method.

Methods

Reference-counted, clone-on-write collection of synchronous and asynchronous methods.

NoopIdProvider

No-op implementation to be used for servers that don’t support subscriptions.

PendingSubscriptionAcceptError

The error returned while accepting a subscription.

PendingSubscriptionSink

Represents a single subscription that is waiting to be accepted or rejected.

RandomIntegerIdProvider

Generates random integers as subscription ID.

RandomStringIdProvider

Generates random strings of length len as subscription ID.

RpcModule

Sets of JSON-RPC methods can be organized into “module“s that are in turn registered on the server or, alternatively, merged with other modules to construct a cohesive API. RpcModule wraps an additional context argument that can be used to access data during call execution.

Server

JSON RPC server.

ServerBuilder

Builder to configure and create a JSON-RPC server

ServerHandle

Server handle.

Subscription

Wrapper struct that maintains a subscription “mainly” for testing.

SubscriptionKey

Represent a unique subscription entry based on SubscriptionId and ConnectionId.

SubscriptionMessage

Subscription message.

SubscriptionSink

Represents a single subscription that hasn’t been processed yet.

SubscriptionState

Helper struct to manage subscriptions.

Enums

BatchRequestConfig

Configuration for batch request handling.

CallOrSubscription

This represent a response to a RPC call and Subscribe calls are handled differently because we want to prevent subscriptions to start before the actual subscription call has been answered.

MethodCallback

Callback wrapper that can be either sync or async.

MethodResult

Result of a method, either direct value or a future of one.

SendTimeoutError

Error that may occur during crate::server::MethodSink::send_timeout or crate::server::SubscriptionSink::send_timeout.

SubscriptionCloseResponse

Represents what action that will sent when a subscription callback returns.

SubscriptionMessageInner

A complete subscription message or partial subscription message.

TrySendError

Error that may occur during crate::server::MethodSink::try_send or crate::server::SubscriptionSink::try_send.

Traits

IdProvider

Trait to generate subscription IDs.

IntoResponse

Something that can be converted into a JSON-RPC method call response.

IntoSubscriptionCloseResponse

Convert something into a subscription close notification before a subscription is terminated.

Type Aliases

AsyncMethod

Similar to SyncMethod, but represents an asynchronous handler.

ConnectionId

Connection ID, used for stateful protocol such as WebSockets. For stateless protocols such as http it’s unused, so feel free to set it some hardcoded value.

MaxResponseSize

Max response size.

RawRpcResponse

Raw response from an RPC A tuple containing:

Subscribers

Type-alias for subscribers.

SubscriptionMethod

Method callback for subscriptions.

SubscriptionPermit

Subscription permit.

SyncMethod

A MethodCallback is an RPC endpoint, callable with a standard JSON-RPC request, implemented as a function pointer to a Fn function taking four arguments: the id, params, a channel the function uses to communicate the result (or error) back to jsonrpsee, and the connection ID (useful for the websocket transport).