Module jsonrpsee_server::helpers

Helpers.

Structs

BatchResponseBuilder

Builder to build a BatchResponse.

BoundedWriter

Bounded writer that allows writing at most max_len bytes.

MethodResponse

Represent the response to a method call.

MethodSink

Sink that is used to send back the result to the server for a specific method.

Enums

MethodResponseResult

Represent the outcome of a method call success or failed.

Functions

batch_response_error

Create a JSON-RPC error response.

prepare_error

Figure out if this is a sufficiently complete request that we can extract an Id out of, or just plain unparseable garbage.