Module jsonrpsee_server::logger

Logger for jsonrpsee servers.

Structs

Body

A stream of Bytes, used when receiving bodies.

Headers

A set of HTTP headers

Params

Parameters sent with an incoming JSON-RPC request.

Enums

MethodKind

The type JSON-RPC v2 call, it can be a subscription, method call or unknown.

TransportProtocol

The transport protocol used to send or receive a call or request.

Traits

Logger

Defines a logger specifically for WebSocket connections with callbacks during the RPC request life-cycle. The primary use case for this is to collect timings for a larger metrics collection solution.

Type Aliases

HttpRequest

HTTP request.