jsonrpsee_server

Function serve

pub async fn serve<S, B, I>(io: I, service: S) -> Result<(), BoxError>
where
    S: Service<Request<Incoming>, Response = Response<B>> + Clone + Send + 'static,
    S::Future: Send,
    S::Response: Send,
    S::Error: Into<BoxError>,
    B: Body<Data = Bytes> + Send + 'static,
    B::Error: Into<BoxError>,
    I: AsyncRead + AsyncWrite + Send + Unpin + 'static,

Serve a service over a TCP connection without graceful shutdown. This means that pending requests will be dropped when the server is stopped.

If you want to gracefully shutdown the server, use serve_with_graceful_shutdown instead.