jsonrpsee_server::http

Module response

HTTP response helpers.

Functions

denied

Create a response for when the server denied the request.

host_not_allowed

Create a text/plain response for not allowed hosts.

internal_error

Create a response for json internal error.

malformed

Create a json response for empty or malformed requests (400)

method_not_allowed

Create a text/plain response for disallowed method used.

ok_response

Create a valid JSON response.

too_large

Create a json response for oversized requests (413)

too_many_requests

Create a response for when the server is busy and can’t accept more requests.

unsupported_content_type

Create a response for unsupported content type.