Trait jsonrpsee_server::middleware::rpc::RpcServiceT

pub trait RpcServiceT<'a> {
    type Future: Future<Output = MethodResponse> + Send;

    // Required method
    fn call(&self, request: Request<'a>) -> Self::Future;
}

Similar to the tower::Service but specific for jsonrpsee and doesn’t requires &mut self for performance reasons.

Required Associated Types

type Future: Future<Output = MethodResponse> + Send

The future response value.

Required Methods

fn call(&self, request: Request<'a>) -> Self::Future

Process a single JSON-RPC call it may be a subscription or regular call. In this interface they are treated in the same way but it’s possible to distinguish those based on the MethodResponse.